requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.1
django-appconf==0.6
django-compressor==1.4
django-debug-toolbar==1.2.2
django-filter==0.11.0
django-toolbelt==0.0.1
djangorestframework==3.0.0
drf-nested-routers==0.9.0
gunicorn==19.1.1
Markdown==2.6.2
psycopg2==2.5.4
simplejson==3.8.0
six==1.8.0
sqlparse==0.1.13
static3==0.5.1
wheel==0.24.0
